Restructure code to use vectors, write single words code
Vectors were implemented to contain possible strings that can be added to a password attempt. This structure was chosen for scalability and the ability to enter "00" through "09" as strings more easily. Vectors were chosen over arrays in order to allow for the transition to a more fully funcitoning program which can accept parameter input.

Code for multiple word passwords needs to be written, and a structure should be developed to allow the program to loop through all n possible positions.